Got dynamic page loading to work
